put the dict generation for chain in the proper place
