+++ mcm/rest_api/ChainedRequestActions.py
@@ -15,6 +15,7 @@ from tools.json import threaded_loads
 import traceback
 from tools.locker import locker
 from tools.settings import settings
+from tools.request_to_wma import request_to_wma
 
 """
 ## import the rest api for wilde request search and dereference to chained_requests using member_of_chain
@@ -641,159 +642,8 @@ class TaskChainDict(RESTResource):
         """
         if not len(args):
             return dumps({})
-
         arg0 = args[0]
-        crdb = database('chained_requests')
-        rdb = database('requests')
-        def request_to_tasks( r , base, depend):
-            events_per_lumi = settings().get_value('events_per_lumi')
-            ts=[]
-            for si in range(len(r.get_attribute('sequences'))):
-                task_dict={"TaskName": "%s_%d"%( r.get_attribute('prepid'), si),
-                           "KeepOutput" : True,
-                           "ConfigCacheID" : None,
-                           "GlobalTag" : r.get_attribute('sequences')[si]['conditions'],
-                           "CMSSWVersion" : r.get_attribute('cmssw_release'),
-                           "ScramArch": r.get_scram_arch(),
-                           "PrimaryDataset" : r.get_attribute('dataset_name'),
-                           "AcquisitionEra" : r.get_attribute('member_of_campaign'),
-                           "ProcessingString" : r.get_processing_string(si),
-                           "ProcessingVersion" : r.get_attribute('version'),
-                           "TimePerEvent" : r.get_attribute("time_event"),
-                           "SizePerEvent" : r.get_attribute('size_event'),
-                           "Memory" : r.get_attribute('memory'),
-                           "FilterEfficiency" : r.get_efficiency()
-                           }
-                
-                if len(r.get_attribute('config_id'))>si:
-                    task_dict["ConfigCacheID"] = r.get_attribute('config_id')[si]
-
-                if len(r.get_attribute('keep_output'))>si:
-                    task_dict["KeepOutput"] = r.get_attribute('keep_output')[si]
-
-                if r.get_attribute('pileup_dataset_name'):
-                    task_dict["MCPileup"] = r.get_attribute('pileup_dataset_name')
-                    
-                if si==0:
-                    if base:
-                        task_dict.update({"SplittingAlgo"  : "EventBased",
-                                          "RequestNumEvents" : r.get_attribute('total_events'),
-                                          "Seeding" : "AutomaticSeeding",
-                                          "EventsPerLumi" : events_per_lumi,
-                                          "LheInputFiles" : r.get_attribute('mcdb_id')>0
-                                          })
-                        ## temporary work-around for request manager not creating enough jobs
-                        ## https://github.com/dmwm/WMCore/issues/5336
-                        ## inflate requestnumevents by the efficiency to create enough output
-                        max_forward_eff = r.get_forward_efficiency()
-                        task_dict["EventsPerLumi"] /= task_dict["FilterEfficiency"] #should stay nevertheless as it's in wmcontrol for now
-                        task_dict["EventsPerLumi"] /= max_forward_eff #this does not take its own efficiency
-                        task_dict["RequestNumEvents"] /= task_dict["FilterEfficiency"] #should disappear with the above getting resolved
-
-                    else:
-                        if depend:
-                            task_dict.update({"SplittingAlgo"  : "EventAwareLumiBased",
-                                              "InputFromOutputModule" : None,
-                                              "InputTask" : None})
-                        else:
-                            task_dict.update({"SplittingAlgo"  : "EventAwareLumiBased",
-                                              "InputDataset" : r.get_attribute('input_dataset')})
-                else:
-                    task_dict.update({"SplittingAlgo"  : "EventAwareLumiBased",
-                                      "InputFromOutputModule" : ts[-1]['output_'],
-                                      "InputTask" : ts[-1]['TaskName']})
-                task_dict['output_'] = "%soutput"%(r.get_attribute('sequences')[si]['eventcontent'][0])
-                task_dict['priority_'] = r.get_attribute('priority')
-                task_dict['request_type_'] = r.get_wmagent_type()
-                ts.append(task_dict)    
-            return ts
-        
-        if not crdb.document_exists( arg0 ):
-            ## it's a request actually, pick up all chains containing it
-            mcm_r = rdb.get( arg0 )
-            #mcm_crs = crdb.query(query="root_request==%s"% arg0) ## not only when its the root of
-            mcm_crs = crdb.query(query="contains==%s"% arg0)
-            task_name = 'task_'+arg0
-        else:
-            mcm_crs = [crdb.get( arg0 )]
-            task_name = arg0
-
-        if len(mcm_crs)==0:  return dumps({})
-            
-        tasktree = {}
-        ignore_status=False
-        
-        if 'scratch' in argv:
-            ignore_status = True
-        veto_point=None
-        if 'upto' in argv:
-            veto_point=int(argv['upto'])
-
-        for mcm_cr in mcm_crs:
-            starting_point=mcm_cr['step']
-            if ignore_status: starting_point=0
-            for (ir,r) in enumerate(mcm_cr['chain']):
-                if (ir<starting_point) : 
-                    continue ## ad no task for things before what is already done
-                if veto_point and (ir>veto_point):
-                    continue
-                mcm_r = request( rdb.get( r ) )
-                if mcm_r.get_attribute('status')=='done' and not ignore_status:
-                    continue
-
-                if not r in tasktree:
-                    tasktree[r] = { 
-                        'next' : [],
-                        'dict' : [],
-                        'rank' : ir
-                        }
-                base=(ir==0) ## there is only one that needs to start from scratch
-                depend=(ir>starting_point) ## all the ones later than the starting point depend on a previous task
-                if ir<(len(mcm_cr['chain'])-1):
-                    tasktree[r]['next'].append( mcm_cr['chain'][ir+1])
-
-                tasktree[r]['dict'] = request_to_tasks( mcm_r, base, depend)
-
-        for (r,item) in tasktree.items():
-            for n in item['next']:
-                tasktree[n]['dict'][0].update({"InputFromOutputModule" : item['dict'][-1]['output_'],
-                                                       "InputTask" : item['dict'][-1]['TaskName']})
-        
-
-        wma={
-            "RequestType" : "TaskChain",
-            "inputMode" : "couchDB",
-            "Group" : "ppd",
-            "Requestor": "pdmvserv",
-            "OpenRunningTimeout" : 43200,
-            "TaskChain" : 0,
-            "ProcessingVersion": 1,
-            "RequestPriority" : 0,
-            "SubRequestType" : "MC" 
-            }
-
-        task=1
-        for (r,item) in sorted(tasktree.items(), key=lambda d: d[1]['rank']):
-            for d in item['dict']:
-                if d['priority_'] > wma['RequestPriority']:  wma['RequestPriority'] = d['priority_']
-                if d['request_type_'] in ['ReDigi']:  wma['SubRequestType'] = 'ReDigi'
-                for k in d.keys():
-                    if k.endswith('_'):
-                        d.pop(k)
-                wma['Task%d'%task] = d
-                task+=1
-        wma['TaskChain'] = task-1
-
-        if wma['TaskChain'] == 0:
-            return dumps({})    
-        
-        for item in ['CMSSWVersion','ScramArch','TimePerEvent','SizePerEvent','GlobalTag','Memory']:
-            wma[item] = wma['Task%d'% wma['TaskChain']][item]
-
-        wma['Campaign' ] = wma['Task1']['AcquisitionEra']
-        wma['PrepID' ] = task_name
-        wma['RequestString' ] = wma['PrepID']
-        return dumps(wma)
+        return dumps( request_to_wma.get_chain_dict( arg0 ))
             
 
 class GetSetupForChains(RESTResource):
